related_flask/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.27
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-Cors==3.0.10
psycopg2-binary==2.9
dataclasses==0.6
kafka-python==2.0.2
strsimpy
nltk==3.6.5